rename: 요구사항에 맞춰 racingCarGame 폴더 이름 변경
