+++ src/App.js
@@ -1,4 +1,4 @@
-import RacingCarGame from "./RacingCarGame/RacingCarGame.js";
+import RacingCarGame from "./domain/RacingCarGame.js";
 
 class App {
   async play() {
